requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asyncpg==0.30.0
attrs==25.4.0
backports.zstd==1.0.0
bcrypt==5.0.0
brotli==1.2.0
cffi==2.0.0
click==8.3.1
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
propcache==0.4.1
pycares==4.11.0
pycparser==2.23
pytokens==0.3.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
yarl==1.22.0
pydantic==2.9.2
PyJWT==2.10.1